modify eval pertial evaluate to use new pairing index function
