+++ polynomial/src/multilinear/mod.rs
@@ -2,4 +2,3 @@ mod boolean_hypercube;
 pub mod coefficient_form;
 pub mod evaluation_form;
 pub mod pairing_index;
-pub mod pairing_index_2;
